[Minor]  Extracting ClusterMarker
Minor - Extracting ClusterMarker to an independent file
